% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissmacrodata.R
\name{swissmacrodata}
\alias{swissmacrodata}
\title{Get various swiss macro data}
\usage{
swissmacrodata()
}
\value{

}
\description{
The 'SwissMacroData' package provides a convienent way to access a large dataset of macroeconomic data for Switzerland. It is open-source and welcomes contributions from anyone.
}
\examples{
swissmacrodata()
}
